+++ docs/src/py/notebook.py
@@ -569,8 +569,10 @@ def _color_command_console(colors, gamut=WEBSPACE):
     bar = False
     values = []
     for item in colors:
-        if isinstance(item, (HtmlGradient, Steps)):
-            current = total = percent = 0
+        is_grad = isinstance(item, HtmlGradient)
+        is_steps = isinstance(item, Steps)
+        if is_grad or is_steps:
+            current = total = percent = last = 0
             if isinstance(item, Steps):
                 total = len(item)
                 percent = 100 / total
@@ -584,15 +586,20 @@ def _color_command_console(colors, gamut=WEBSPACE):
             stops = []
             for e, color in enumerate(item):
                 color.fit(gamut)
+                color_str = color.convert(gamut).to_string()
                 if current:
-                    stops.append('{}'.format(color.convert(gamut).to_string()))
-                    stops.append('{}'.format(color.convert(gamut).to_string()))
+                    if is_steps:
+                        stops.append('{} {}%'.format(color_str, str(last)))
+                        stops.append('{} {}%'.format(color_str, str(current)))
+                    else:
+                        stops.append(color_str)
+                    last = current
                     if e < (total - 1):
                         current += percent
                     else:
                         current = 100
                 else:
-                    stops.append(color.convert(gamut).to_string())
+                    stops.append(color_str)
             if not stops:
                 stops.extend(['transparent'] * 2)
             if len(stops) == 1: